Added update-cache.command
Update cache command will check an MD5 against a locally downloaded
cache. If the cache doesn't match, a new file will be downloaded. If
that file doesn't match, we fail to update.

Also checks for a VERSION.txt in the root directory of the
$HOME/.worlds_of_minecraft_cache path to match. If it doesn't, the
cache is extracted there. If it still doesn't match, the update fails.